fix: FoodTagListener setCursor/setCurrentItem, NPE guard, CharLevel MEDIUM, StoveBlock hand filter, FuelEntry int, rebuildBoardDisplays pos

+++ src/main/java/io/github/thebusybiscuit/exoticgarden/cooking/state/FuelEntry.java
@@ -2,9 +2,9 @@ package io.github.thebusybiscuit.exoticgarden.cooking.state;
 
 public class FuelEntry {
     public String fuelId;
-    public double ticksRemaining;
+    public int ticksRemaining;
 
-    public FuelEntry(String fuelId, double ticksRemaining) {
+    public FuelEntry(String fuelId, int ticksRemaining) {
         this.fuelId = fuelId;
         this.ticksRemaining = ticksRemaining;
     }
